+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## UNRELEASED
 
 - `Jsonnet.prototype.{extString,extCode,tlaString,tlaCode}` now accept an object to assign multiple variables at once.
+- Introduce `Jsonnet.prototype.trailingNewline` to control whether the output will have a trailing newline.
 
 ## v3.1.1 (2026-04-02)